+++ _extensions/hebstr-doc/_extensions/mcanouil/code-window/_vendor/quarto-lua-modules/pandoc-helpers.lua
@@ -1,9 +1,8 @@
 --- MC Pandoc Helpers - Pandoc element construction and format detection for Quarto Lua filters and shortcodes
---- @module pandoc_helpers
+--- @module "pandoc_helpers"
 --- @license MIT
 --- @copyright 2026 Mickaël Canouil
 --- @author Mickaël Canouil
---- @version 1.0.0
 
 local M = {}